First step in making this more generic - not only modules, but addons in general. No need to maintain multiple composer plugins. NOT TESTED YET
