+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "impresscms/composer-module-installer-plugin",
-    "description": "This plugin allows the installation of ImpressCMS modules from composer",
+    "description": "This plugin allows the installation of ImpressCMS addons from composer",
     "minimum-stability": "dev",
     "license": "MIT",
     "authors": [
@@ -14,8 +14,8 @@
         "psr-4": {"ImpressCMS\\Composer\\": "src/"}
     },
     "extra": {
-        "class": "ImpressCMS\\Composer\\ModuleInstallerPlugin"
-    },
+        "class": "ImpressCMS\\Composer\\AddonInstallerPlugin"
+        },
     "require": {
         "composer-plugin-api": "1.0.0"
     }
