--- src/ThemeInstallerPlugin.php
@@ -1,29 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: fiammy
- * Date: 15/10/15
- *
- */
-
-namespace ImpressCMS\Composer;
-
-use Composer\Composer;
-use Composer\IO\IOInterface;
-use Composer\Plugin\PluginInterface;
-
-class ThemeInstallerPlugin implements PluginInterface {
-    /**
-     * activate - add our installer to composer
-     *
-     * @param Composer $composer composer instance
-     * @param IOInterface $io composer i/o
-     *
-     * @return void
-     */
-    public function activate(Composer $composer, IOInterface $io)
-    {
-        $installer = new ThemeInstaller($io, $composer);
-        $composer->getInstallationManager()->addInstaller($installer);
-    }
-}
